% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{abc2index}
\alias{abc2index}
\title{Get the numbers (indices) corresponding to alphabetical colnames}
\usage{
abc2index(abc)
}
\arguments{
\item{abc}{a character vector to be converted to numbers (indices).}
}
\value{
an integer vector
}
\description{
This function supports nchar up to 4, i.e. from "A" to "ZZZZ".
}
\examples{
abc2index(letters)         # 1–26
abc2index(c("a", NA, "a")) # 1 NA 1
\dontrun{
 abc2index("zzzza")         # Error
}
}
