% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfn.r
\name{expfn}
\alias{expfn}
\title{Convert numbers as expressions of exponential format}
\usage{
expfn(
  x,
  digits = 3,
  scale = 1,
  prefix = "",
  suffix = "",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\arguments{
\item{x}{numeric vector.}

\item{digits}{a number of digits just before '×'. Default: 3}

\item{scale}{a number to multiply x. Default: 1.}

\item{prefix}{a character string to paste just before x. Default: ''.}

\item{suffix}{a character string to paste just after x. Default: ''.}

\item{decimal.mark}{a character string to use decimal mark, Default: '.'.}

\item{trim}{Logical, if ‘FALSE’, values are right-justified to a common
width (see ‘base::format()’). Default: TRUE.}

\item{...}{extra arguments passed to \code{scientific()}.}
}
\value{
An expression
}
\description{
Modify input numbers to expression class of exponential format like \code{expression(1.35~'×'~10^{4})}.
}
\details{
DETAILS
}
\examples{
expfn(c(1.2e8, 1.03e8, -Inf))
expfn(-Inf)
}
\seealso{
\code{\link[stringr]{str_extract}}
\code{\link[scales]{scientific}}
}
