% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{get_styles}
\alias{get_styles}
\title{Get font names corresponding to four style names for a font family}
\usage{
get_styles(data)
}
\arguments{
\item{data}{\code{data.frame} alike, output of font list from \code{systemfonts:: system_fonts()} or its subset, which includes six columns: \code{name}, \code{family},
\code{style}, \code{weight}, \code{width}, and \code{italic}/ \code{data.frame} 相当のオブジェクトで,
\code{systemfonts::system_fonts()} のから特定のファミリだけ取り出したもので,
\code{name}, \code{family}, \code{style}, \code{weight}, \code{width}, \code{italic} 列を持つもの.}
}
\value{
character vector of length 4: It's elements are names of plain,
bold, italic, and bold italic font./長さ4の文字列ベクトル. 要素は順に標準,
ボールド, イタリック, ボールドイタリック体のpostscript名.
}
\description{
\code{systemfonts::system_fonts()} などの結果から4書体のフォントを取
得する. ただしうまくいくとは限らない.
}
\details{
The original function was transferred from \url{https://github.com/Gedevan-Aleksizde/fontregisterer} (Version 0.3) accessed on 2023/06/02 (Fri.).
}
