% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{index2abc}
\alias{index2abc}
\title{FUNCTION_TITLE}
\usage{
index2abc(num, nchar = NULL)
}
\arguments{
\item{num}{PARAM_DESCRIPTION}

\item{nchar}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
