% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurai_ja.r
\name{kurai_ja}
\alias{kurai_ja}
\title{Japanese scale units by 'Man-shin-hou'}
\usage{
kurai_ja(num, remove_ichi = TRUE, scipen = 108)
}
\arguments{
\item{num}{A numeric value or a numeric vector.
Smaller numbers after decimal point are rounded.}

\item{remove_ichi}{Not print '一' when TRUE (default).}

\item{scipen}{An integer. A penalty to be applied when deciding to print
numeric values in fixed or exponential notation. Default: 108.}
}
\value{
Japanese scale units by Kanji
}
\description{
入力された数値に関して万進法 (中数) での日本語単位を返す関数.
}
\examples{
if(interactive()){
 kurai_ja(c(1, 24, 100, 1000, 11111, 6e18, 1e72, -123489)) 
 }
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[stringr]{str_remove}}
}
