% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfn.r
\name{label_expfn}
\alias{label_expfn}
\title{Label exponential format}
\usage{
label_expfn(
  digits = 3,
  scale = 1,
  prefix = "",
  suffix = "",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\arguments{
\item{digits}{a number of digits before \verb{\\\\u00d7}. Default: 3}

\item{scale}{a number to multiply x. Default: 1.}

\item{prefix}{a character string to paste just before x. Default: ''.}

\item{suffix}{a character string to paste just after x. Default: ''.}

\item{decimal.mark}{a character string to use decimal mark, Default: '.'.}

\item{trim}{Logical, if ‘FALSE’, values are right-justified to a common
width (see ‘base::format()’). Default: TRUE.}

\item{...}{extra arguments passed to \code{scientific()}.}
}
\value{
A function
}
\description{
Return functions to display exponential format expression where
\code{expfn} works internally.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
