% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixfonts.r
\name{label_mixfonts}
\alias{label_mixfonts}
\title{Label mixed-fonts strings in ggplot}
\usage{
label_mixfonts(asciifont = "Arial", mbytefont = "MS Gothic")
}
\arguments{
\item{asciifont}{ASCII font family, Default: 'Arial'}

\item{mbytefont}{Multi byte font family, Default: 'MS Gothic'}
}
\value{
\code{mixfonts} function
}
\description{
Label mixed-fonts strings in ggplot
}
\details{
See vignettees("mixfonts")
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
