% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl2blank.r
\name{label_repl2blank}
\alias{label_repl2blank}
\title{軸ラベルを一定間隔でブランクに置き換える関数を返す}
\usage{
label_repl2blank(
  nskip = 1,
  first = FALSE,
  last = FALSE,
  inverse = FALSE,
  ptime = FALSE,
  ftime = "\%Y-\%m-\%d"
)
}
\arguments{
\item{nskip}{何個置きに空白文字にするか, Default: 1.}

\item{first}{1個目の値を空白にする, Default: FALSE. TRUEで強制的に空白にする. inverse = TRUE の時はTRUEで強制的に値を与える.}

\item{last}{最後の値を空白にする, Default: FALSE. TRUEで強制的に空白にする. inverse = TRUE の時はTRUEで強制的に値を与える.}

\item{inverse}{空白と非空白を逆転させる, Default: FALSE}

\item{ptime}{Parse date-time type vector, Default: FALSE}

\item{ftime}{Format of date-time, Default: '\%Y-\%m-\%d'}
}
\value{
A function
}
\description{
ggplot作図の際に軸ラベルの一部を空白にすることで minor break のようにみせかける.
}
\details{
例をご覧ください.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{force_all}
