% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{make_abcbase}
\alias{make_abcbase}
\title{FUNCTION_TITLE}
\usage{
make_abcbase(nchar = NULL)
}
\arguments{
\item{nchar}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
}
