% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{makeXlColAlphabet}
\alias{makeXlColAlphabet}
\title{Make Microsoft Excel's alphabet column names}
\usage{
makeXlColAlphabet(nchar = 2L, len = NULL, start = 1L)
}
\arguments{
\item{nchar}{A number of combined characters, Default: 2}

\item{len}{Length (size) of the output vector, Default: NULL}

\item{start}{Index at start, Default: 1. When 'start' = 1 (default) is "A"; 'start' = 2 is "B", for example.}
}
\value{
A character vector.
}
\description{
Genarate a character vector such as "A" "B" ... "AA" "AB" ... "ZZZ".
}
\details{
No description.
}
\examples{
\dontrun{
if(interactive()){
makeXlColAlphabet(nchar = 1)
makeXlColAlphabet(len = 10, start = 3)
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
\code{\link[dplyr]{case_when}}
}
