% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixfont4ggplot.r
\name{mixFonts}
\alias{mixFonts}
\title{To display two font families in a single character string on ggplot title and labels}
\usage{
mixFonts(string, asciifont = "Arial", mbytefont = "MS Gothic")
}
\arguments{
\item{string}{A character string to be converted}

\item{asciifont}{ASCII font family, Default: 'Arial'}

\item{mbytefont}{Multi byte font family, Default: 'MS Gothic'}
}
\value{
An HTML <span> tag string.
}
\description{
This function has been made to cope with VERY VERY specific
demand in Japanese domestic Stock assessment.
}
\details{
See also \code{ggtext} package
}
\examples{
\dontrun{
if(interactive()){
 # EXAMPLE1
 mixFonts("<br>シラスを除く<br>カタクチイワシ0歳魚")
 }
}
}
\seealso{
\code{\link[stringr]{str_split}}
\code{\link[foreach]{foreach}}
\code{\link[htmltools]{builder}}
}
