% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfkc_all.r
\name{nfkc_all}
\alias{nfkc_all}
\title{Unicode normalization of colnames and contents of data.frame with
compatibility equivalence}
\usage{
nfkc_all(.data)
}
\arguments{
\item{.data}{data.frame-like object (tibble, data.table, etc.)}
}
\value{
object of the same class as .data
}
\description{
This is a dplyr-dependent preprocessing function
}
\details{
""
}
\examples{
 # data.frame
 df <- data.frame(`列１（トン）` = 1, `列2 (kg) ` = 2)
 nfkc_all(df)
 # tibble
 tbl <- tibble::tibble(`列１（トン）` = 1, `列2 (kg) ` = 2)
 nfkc_all(tbl)
 # data.table
 dt <- data.table::data.table(`列１（トン）` = 1, `列2 (kg) ` = 2)
 nfkc_all(dt)
}
