% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_all_fonts}
\alias{register_all_fonts}
\title{Register all fonts}
\usage{
register_all_fonts()
}
\description{
register all system fonts obtained from \code{systemfonts} package
by \code{windowsFonts} or \code{quartzFonts}/ \code{systemfonts} で取得したフォント名を全て
OS固有のデバイス (\code{windowsFonts}, \code{quartzFonts}) に登録する.
}
\details{
The original function was transferred from \url{https://github.com/Gedevan-Aleksizde/fontregisterer} (Version 0.3) accessed on 2023/06/02 (Fri.).
}
