% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{shift_abc}
\alias{shift_abc}
\title{Shift n from given alphabetical colname}
\usage{
shift_abc(abc, n_shift = 0L)
}
\arguments{
\item{abc}{a character vector to be shifted.}

\item{n_shift}{numbers to shift, Default: 0}
}
\value{
a character vector
}
\description{
This function is useful for output with openxlsx package.
}
\examples{
shift_abc("a", n_shift = 3)        # D
shift_abc(letters, n_shift = 3)    # D–AC
shift_abc(letters, n_shift = 1:26) # B–AZ. skipping ahead one by one.
}
