% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_resc.r
\name{va_resc}
\alias{va_resc}
\title{2軸グラフ用の数値調整関数}
\usage{
va_resc(v, lim1 = NULL, lim2 = NULL, scale_which = "var", return_lims = FALSE)
}
\arguments{
\item{v}{Values to plot. 描画したい数値のベクトル.}

\item{lim1}{Limit(s) of 1st (left-hand) y-axis. 左軸に対してplotしたい変数の範囲 (要素数2) またはvector (要素数3以上).
vectorを指定した場合は range(lim1) となる. また, 正数であれば最大値のみの指定も可能 (最小値は0となる).
NULLの場合は range(v) となる. Default: NULL}

\item{lim2}{Limit(s) of 2nd (right-hand) y-axis. 右軸に対してplotしたい変数の範囲 (要素数2) またはvector (要素数3以上).
vectorを指定した場合は range(lim2) となる. また, 正数であれば最大値のみの指定も可能 (最小値は0となる).
NULLの場合は range(v) となる. Default: NULL}

\item{scale_which}{A string. 右軸 (2nd y-axis) に対して描画したい変数を左軸にスケーリングする場合は\code{var},
右軸の目盛りを変換する場合は\code{axis}. Default: "var".}

\item{return_lims}{Logical. 変換後の数値の代わりに範囲を標準 (message) 出力する, Default: FALSE}
}
\value{
変換後の数値
}
\description{
2軸グラフを描く時は左軸に値をスケーリングし, 右軸は目盛り数値を調整する.
}
\details{
See example.
}
\examples{
\dontrun{
if(interactive()){
#EXAMPLE1
lh <- range(faithful$waiting)
rh <- range(faithful$eruptions)

faithful \%>\%
    tidyr::pivot_longer(cols = dplyr::everything()) \%>\%
    dplyr::mutate(val = ifelse(name == "eruptions", 
                               va_resc(value, lh, rh, "var"), value)) \%>\%
    ggplot(aes(x = as.integer(rownames(.)), group = name)) +
    geom_line(aes(y = val, color = name), size = 0.3) +
    scale_y_continuous(name = "waiting",
                       sec.axis = sec_axis(~ va_resc(., lh, rh, "axis"),
                                           name = "eruptions"))
 }
}
}
\references{
\url{https://blog.statsbeginner.net/entry/2020/09/08/181536}
\url{https://knknkn.hatenablog.com/entry/2019/02/27/204913"}
\url{https://notchained.hatenablog.com/entry/2016/10/02/011204"}
}
