% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{abc2index}
\alias{abc2index}
\title{FUNCTION_TITLE}
\usage{
abc2index(abc, nchar = NULL)
}
\arguments{
\item{abc}{PARAM_DESCRIPTION}

\item{nchar}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
abc2index(letters)
abc2index(c("a", "a"))
 }
}
}
\seealso{
\code{\link[stringr]{str_subset}}
}
