% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{conv_dm2dd}
\alias{conv_dm2dd}
\title{Convert DM coordinates values (sexagesimal) to DD (decimal degree)}
\usage{
conv_dm2dd(dm60, original_value = TRUE, num_as_dm = FALSE)
}
\arguments{
\item{dm60}{sexagesimal numbers of character / numeric class.}

\item{sep_marks}{a separation character between D and M part. Default: c(".",
"-", "_", "u00b0")}

\item{out_original}{include original values to output. Default: FALSE.}

\item{out_original_D}{D (degrees) part on console to make sure the
conversion is correct. Default: FALSE.}

\item{out_original_M}{print M (minutes) part on console to make sure the
conversion is correct. Default: FALSE.}
}
\value{
a vector of numeric class (default) or a data.frame if any output
options are TRUE.
}
\description{
This function is used to convert latitude and longitude from
degrees, minutes format to decimal degree format, which is useful for map
drawing.
}
\details{
See example
}
\examples{
if(interactive()){
 dmx <- c("35", "35.30", "35.3", "35.03", "35. 3", "35.    3", "30.30.750")
 conv_dm2dd(dmx)
 dmx2 <- c("35°30′", "30°30′750″", "30°30.750′", "-35-30", "-30-  3")
 conv_dm2dd(dmx2)
 conv_dm2dd(dmx2, original_value = FALSE)

 dmnum <- c(34.30, 34.3, 34.03, 34, -120.5, -30.8)
 conv_dm2dd(dmnum)
 conv_dm2dd(dmnum, num_as_dm = TRUE)
 
 conv_dm2dd("-135°30′")
}
}
