% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_panel_ranges.r
\name{coord_panel_ranges}
\alias{coord_panel_ranges}
\title{facet作図後にパネルごとに異なるcoordinateを適用する関数}
\usage{
coord_panel_ranges(panel_ranges, expand = TRUE, default = FALSE, clip = "on")
}
\arguments{
\item{panel_ranges}{Limits of axes or an axis.}

\item{expand}{Add margin to each axes or an axis, Default: TRUE}

\item{default}{PARAM_DESCRIPTION, Default: FALSE}

\item{clip}{PARAM_DESCRIPTION, Default: 'on'}
}
\value{
ggproto object
}
\description{
個々のパネルの座標を調整する.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
library(frabento)
library(ggplot2)
library(dplyr)
library(patchwork)
 #EXAMPLE1
theme_set(theme_linedraw(base_family = "Helvetica", base_line_size = 0.3) +
          theme(aspect.ratio = 1/1))

# base plot
g <- iris \%>\%
    ggplot(aes(x = Sepal.Length, y = Sepal.Width)) +
    geom_point(aes(color = Species)) +
    geom_rect(data = .  \%>\% dplyr::filter(Species == "versicolor"),
              xmin = 5.5, xmax = 6.5, ymin = 2.5, ymax = 3,
              color = "red", size = 1, fill = NA) +
    facet_wrap(~ Species, scale = "free", ncol = 2, dir = "h") +
    theme(legend.position = c(0.75, 0.25))

# zoom in versicolor panel (this is second panel)
gz <- g + coord_panel_ranges(panel_ranges = list(
    list(NULL),
    list(x = c(5.4, 6.6), y = c(2.4, 3.1)),
    list(NULL)
    ))

# patchwork
g | gz

 }
}
}
\references{
\url{https://andburch.github.io/ggplot_facets/}
\url{https://www.javaer101.com/article/3690503.html}
}
\seealso{
\code{\link[ggplot2]{ggproto}}
}
