% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils4ggplot.r
\name{coord_panel_ranges}
\alias{coord_panel_ranges}
\title{facet作図後にパネルごとに異なるcoordinateを適用する関数}
\usage{
coord_panel_ranges(panel_ranges, expand = TRUE, default = FALSE, clip = "on")
}
\arguments{
\item{panel_ranges}{Limits of axes or an axis.}

\item{expand}{Add margin to each axes or an axis, Default: TRUE}

\item{default}{PARAM_DESCRIPTION, Default: FALSE}

\item{clip}{PARAM_DESCRIPTION, Default: 'on'}
}
\value{
ggproto object
}
\description{
facet作図後にパネルごとに異なるcoordinateを適用する関数
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{
\url{https://andburch.github.io/ggplot_facets/}
\url{https://www.javaer101.com/article/3690503.html}
}
\seealso{
\code{\link[ggplot2]{ggproto}}
}
