% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{index2abc}
\alias{index2abc}
\title{Get the alphabetical colnames corresponding to numbers (indices)}
\usage{
index2abc(num, nchar = 4L)
}
\arguments{
\item{num}{an integer vector which is converted to alphabets.}

\item{nchar}{an integer used for creating an alphabetical name base.
Default: 4L.}
}
\value{
a character vector
}
\description{
This function supports nchar up to 4 (ZZZZ), i.e. the upper
limit of number is 475,254.
}
\examples{
index2abc(num = 5)        # "E"
index2abc(num = 10^(1:5)) # "J" "CV" "ALL" "NTP" "EQXD"
\dontrun{
 index2abc(num = 500000)   # Error. due to exceeding sum(26^(1:4))
}
}
