% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurai_ja.r
\name{kurai_ja}
\alias{kurai_ja}
\title{Japanese scale units by 'Man-shin-hou'}
\usage{
kurai_ja(num, remove_ichi = TRUE, scipen = 108)
}
\arguments{
\item{num}{A numeric value or a numeric vector.
Smaller numbers after decimal point are rounded.}

\item{remove_ichi}{Not print '一' when TRUE (default).}

\item{scipen}{An integer. A penalty to be applied when deciding to print
numeric values in fixed or exponential notation. Default: 108.}
}
\value{
A character vector of Kanji.
}
\description{
入力された数値に関して万進法 (中数) での日本語単位を返す関数.
}
\details{
Decimal numbers not yet supported such that 0.1 and 0.00001 will be
converted to "" (default) or '一' (if remove_ichi = FALSE).
}
\examples{
kurai_ja(c(1, 23, 456, 7890, 98765, 4e32, 1e+99, -1234567890)) 
kurai_ja(c(0, 1, 0.05), remove_ichi = FALSE) 
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[stringr]{str_remove}}
}
