% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixfonts.r
\name{label_mixfonts}
\alias{label_mixfonts}
\title{Label mixed-fonts HTML strings}
\usage{
label_mixfonts(asciifont = "Arial", mbytefont = "MS Gothic")
}
\arguments{
\item{asciifont}{ASCII (Single byte) font family, Default: 'Arial'}

\item{mbytefont}{Multi byte font family, Default: 'MS Gothic'}
}
\value{
Returns \code{mixfonts()}.
}
\description{
Label mixed-fonts HTML strings
}
\details{
This function is supposed to be used with \code{ggtext::element_markdown() }. See vignettees("mixfonts")
}
\examples{
if(interactive()){
 library(ggplot2)
 library(magrittr)
 library(ggtext)
 library(frabento)
 register_all_fonts()

 set.seed(30)
 data.frame(cpue = c(rnorm(n = 30, mean = 300, sd = 25),
                     rnorm(n = 30, mean = 500, sd = 35)),
            year = rep(seq(1990, length.out = 30, by = 1), times = 2),
            age  = rep(c("0歳魚", "1歳魚+"), each = 30)) \%>\%
     ggplot(aes(x = year, y = cpue, group = age)) +
     geom_path(aes(color = age)) +
     scale_color_discrete(labels = label_mixfonts()) +
     theme(legend.text = element_markdown(color = "blue"),
           legend.position = "right")
}
}
\seealso{
\code{\link[frabento]{mixfonts}}
\code{\link[ggtext]{element_markdown}}
\code{\link[ggtext]{geom_richtext}}
}
