% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl2blank.r
\name{label_repl2blank}
\alias{label_repl2blank}
\title{Label character vector with blanks at a certain interval}
\usage{
label_repl2blank(
  nskip = 1,
  shift = 0,
  first = FALSE,
  last = FALSE,
  inverse = FALSE,
  comma = FALSE,
  ptime = FALSE,
  ftime = "\%Y-\%m-\%d"
)
}
\arguments{
\item{nskip}{何個置きに空白文字にするか, Default: 1.}

\item{shift}{blankで置き換える位置を右にずらす. Default: 0.}

\item{first}{1個目の値を空白にする, Default: FALSE. TRUEで強制的に空白にする
. inverse = TRUE の時は逆動作となる.}

\item{last}{最後の値を空白にする, Default: FALSE. TRUEで強制的に空白にする.
inverse = TRUE の時は逆動作となる.}

\item{inverse}{空白と非空白を逆転させる, Default: FALSE}

\item{comma}{3桁ごとにコンマをつける, Default: FALSE}

\item{ptime}{Parse date-time type vector, Default: FALSE}

\item{ftime}{Format of date-time, Default: '\%Y-\%m-\%d'}
}
\value{
A function
}
\description{
ggplot作図の際に軸ラベルの一部を空白にすることで minor break の
ようにみせかける.
}
\details{
例をご覧ください.
}
\examples{
\dontrun{
if(interactive()){
library(dplyr)
library(tibble)
library(ggplot2)

tibble::tibble(year =  -6000:1979, haba = treering) \%>\%
    ggplot(aes(x = year, y = haba)) +
    geom_line(size = 0.2) +
    scale_x_continuous(breaks = seq(-6000, 2000, by = 200),
                       labels = label_repl2blank(nskip = 4))

tibble::tibble(year =  -6000:1979, haba = treering) \%>\%
    ggplot(aes(x = year, y = haba)) +
    geom_line(size = 0.2) +
    scale_x_continuous(breaks = seq(-6000, 2000, by = 200),
                       labels = label_repl2blank(nskip = 4, comma = TRUE))
 }
}
}
