% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{makeXlColAlphabet}
\alias{makeXlColAlphabet}
\title{Make Microsoft Excel's alphabet column names}
\usage{
makeXlColAlphabet(start = NULL, len = NULL, nchar = 2L, print_all = FALSE)
}
\arguments{
\item{start}{Index or character at start, Default: NULL (to which 'A' is automatically
assigned). \code{start} larger than \code{len} will return NA.}

\item{len}{Length (size) of the output vector, Default: NULL}

\item{nchar}{A number of combined characters, Default: 2}

\item{print_all}{displays all characters to the console if TRUE. When FALSE (default), showing only the first and last characters with column numbers.}
}
\value{
A character vector.
}
\description{
Genarate a character vector such as "A" "B" ... "AA" "AB" ... "ZZZ".
}
\details{
No description.
}
\examples{
\dontrun{
if(interactive()){
makeXlColAlphabet(nchar = 1)
makeXlColAlphabet(len = 10, start = 3)
makeXlColAlphabet(len = length(1978:2020), start = "B", print_all = TRUE)
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
\code{\link[dplyr]{case_when}}
}
