% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{makeXlColAlphabet}
\alias{makeXlColAlphabet}
\title{Make Microsoft Excel's alphabetical column names}
\usage{
makeXlColAlphabet(start = NULL, len = NULL, nchar = 2L, print_all = FALSE)
}
\arguments{
\item{start}{Index or character at start, Default: NULL (to which 'A' is automatically
assigned). \code{start} larger than \code{len} will return NA.}

\item{len}{Length (size) of the output vector, Default: NULL}

\item{nchar}{A number of combined characters, Default: 2}

\item{print_all}{displays all characters to the console if TRUE. When FALSE (default), showing only the first and last characters with column numbers.}
}
\value{
A character vector.
}
\description{
Genarate a character vector such as "A" "B" ... "AA" "AB" ... "ZZZ".
}
\details{
No description.
}
\examples{
makeXlColAlphabet()
makeXlColAlphabet(nchar = 1)
makeXlColAlphabet(start = "B", len = length(1978:2020))
makeXlColAlphabet(start = 23, len = 10, print_all = TRUE)
}
\seealso{
\code{\link[frabento]{shift_abc}}
\code{\link[frabento]{make_abcbase}}
\code{\link[frabento]{index2abc}}
\code{\link[frabento]{abc2index}}
}
