% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{makeXlColAlphabet}
\alias{makeXlColAlphabet}
\title{Make Microsoft Excel's alphabet column names}
\usage{
makeXlColAlphabet(nchar = 2L, len = NULL, start = 1L, fl = FALSE)
}
\arguments{
\item{nchar}{A number of combined characters, Default: 2}

\item{len}{Length (size) of the output vector, Default: NULL}

\item{start}{Index or character at start, Default: 1 (is 'A'). \code{start} larger than \code{len} will return NA.}

\item{fl}{show only first and last character, Default: FALSE}
}
\value{
A character vector.
}
\description{
Genarate a character vector such as "A" "B" ... "AA" "AB" ... "ZZZ".
}
\details{
No description.
}
\examples{
\dontrun{
if(interactive()){
makeXlColAlphabet(nchar = 1)
makeXlColAlphabet(len = 10, start = 3)
makeXlColAlphabet(len = length(1978:2020), start = "B", fl = TRUE)
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
\code{\link[dplyr]{case_when}}
}
