% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{makeXlColAlphabet}
\alias{makeXlColAlphabet}
\title{Making Microsoft Excel's alphabet column names}
\usage{
makeXlColAlphabet(nchar = 2L, len = NULL, start = 1L)
}
\arguments{
\item{nchar}{A number of combined characters, Default: 2}

\item{len}{Length (size) of the output vector, Default: 26L + nchar}

\item{start}{Index at start, Default: 1. For example, when 'start' = 1 (default) is "A"; 'start' = 2 is "B".}
}
\value{
A character vector.
}
\description{
Genarate a character vector such as "A" "B" ... "AA" "AB" ... "ZZZ".
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
}
