% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixfonts.r
\name{mixfonts}
\alias{mixfonts}
\title{Make character vector with different two-font families in ggplot}
\usage{
mixfonts(strs, asciifont = "Arial", mbytefont = "MS Gothic")
}
\arguments{
\item{strs}{A vector of character to be converted}

\item{asciifont}{ASCII font family, Default: 'Arial'}

\item{mbytefont}{Multi byte font family, Default: 'MS Gothic'}
}
\value{
Character vector of HTML \emph{span} tag.
}
\description{
Put ASCII font family character(s) and multibyte font family
character(s) in a single string like graph plotted in Excel.
}
\details{
This function is supposed to be used with \code{element_markdown()} in \code{ggtext} package
}
\examples{
\dontrun{
if(interactive()){
 library(ggplot2)
 library(ggtext)

 # EXAMPLE1
 mixfonts(c("シラス込み0歳魚", "\nシラスを除く\nカタクチイワシ0歳魚"))

 # EXAMPLE2
 ggplot(women) +
     geom_point(aes(x = height, y = weight)) +
     labs(x = mixfonts("身長\n(inch)", asciifont = "Times New Roman")) +
     theme_linedraw(base_family = "MS Gothic") +
     theme(axis.title.x.bottom = element_markdown())
 }
}
}
\seealso{
\code{\link[stringr]{str_split}},\code{\link[stringr]{str_replace}}
\code{\link[foreach]{foreach}},\code{\link[foreach]{\%do\%}}
\code{\link[htmltools]{builder}}
}
