% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixfonts.r
\name{mixfonts}
\alias{mixfonts}
\title{Unify ascii and multibyte fonts into a string}
\usage{
mixfonts(strs, asciifont = "Arial", mbytefont = "MS Gothic")
}
\arguments{
\item{strs}{A vector of character to be converted}

\item{asciifont}{ASCII font family, Default: 'Arial'}

\item{mbytefont}{Multi byte font family, Default: 'MS Gothic'}
}
\value{
Character vector of HTML \emph{span} tag.
}
\description{
Put ASCII font family character(s) and multibyte font family
character(s) into a single string for use in ggplot.
}
\details{
This function is supposed to be used with \code{element_markdown()} in \code{ggtext} package
}
\examples{
\dontrun{
if(interactive()){
 mixfonts(c("全0歳魚", "仔魚を含まない\n0歳魚"))
 }
}
}
\seealso{
\code{\link[stringr]{str_split}},\code{\link[stringr]{str_replace}}
\code{\link[htmltools]{builder}}
}
