% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfkc_all.r
\name{nfkc_all}
\alias{nfkc_all}
\title{Unicode normalization of colnames and contents of data.frame with
compatibility equivalence}
\usage{
nfkc_all(
  .data,
  remove_space_colnames = FALSE,
  remove_space_contents = FALSE,
  coln_spc = "[:blank:]",
  cont_spc = "[:blank:]"
)
}
\arguments{
\item{.data}{data.frame-like object (tibble, data.table, etc.)}

\item{remove_space_colnames}{If TRUE, remove all white spaces from all
column names. Default: FALSE.}

\item{remove_space_contents}{If TRUE, remove all white spaces from all
contents. Default: FALSE.}

\item{coln_spc}{Space character to be removed in colnames. Default: \link{:blank:}}

\item{cont_spc}{Space character to be removed in contents. Default: \link{:blank:}}
}
\value{
object of the same class as .data
}
\description{
This is a dplyr-dependent preprocessing function
}
\examples{
 # data.frame
 df <- data.frame(`列１（トン）` = 1, `列2 (kg) ` = 2)
 nfkc_all(df)

 # data.table
 dt <- data.table::data.table(`列１（トン）` = 1.1, `列2 (kg) ` = 2.1)
 nfkc_all(dt)

 # tibble
 tbl <- tibble::tibble(`列１（トン）` = "1.2\r\nトン",
                       `列2 (kg) ` = " 2.2\n kg")
 nfkc_all(tbl)
 # remove white-spaces (blanks) from all column names
 nfkc_all(tbl, remove_space_colnames = TRUE)
 # remove white-spaces (blanks) from all contents
 nfkc_all(tbl, remove_space_contents = TRUE)
 # remove all white-spaces (blanks) from .data
 nfkc_all(tbl, remove_space_colnames = TRUE, remove_space_contents = TRUE)

 # remove all spaces (including new line characters) from all contents
 nfkc_all(tbl, remove_space_contents = TRUE, cont_spc = "[:space:]")
}
