% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl2blank.r
\name{repl2blank}
\alias{repl2blank}
\title{Replace string elements to blank at a certain interval}
\usage{
repl2blank(
  vec,
  nskip = 1,
  shift = 0,
  first = FALSE,
  last = FALSE,
  inverse = FALSE,
  comma = FALSE
)
}
\arguments{
\item{vec}{A character or an numeric vector. ラベルにする文字列/数値ベクトル}

\item{nskip}{何個置きに空白文字にするか, Default: 1.}

\item{shift}{blankで置き換える位置を右にずらす. Default: 0.}

\item{first}{1個目の値を空白にする, Default: FALSE. TRUEで強制的に空白にする
. inverse = TRUE の時は逆動作となる.}

\item{last}{最後の値を空白にする, Default: FALSE. TRUEで強制的に空白にする.
inverse = TRUE の時は逆動作となる.}

\item{inverse}{空白と非空白を逆転させる, Default: FALSE}

\item{comma}{3桁ごとにコンマをつける, Default: FALSE}
}
\value{
A character vector
}
\description{
ggplot作図の際に軸ラベルの一部を空白にすることで minor break の
ようにみせかける.
}
\details{
See example and also vignette("repl2blank")
}
\examples{
repl2blank(1:10)
repl2blank(1:10, nskip = 2)
repl2blank(1:10, nskip = 2, shift = 1)
}
\seealso{
\code{\link[frabento]{label_repl2blank}}
}
