% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXlColAlphabet.r
\name{shift_abc}
\alias{shift_abc}
\title{FUNCTION_TITLE}
\usage{
shift_abc(abc, n_shift = 0L, nchar = NULL)
}
\arguments{
\item{abc}{PARAM_DESCRIPTION}

\item{n_shift}{PARAM_DESCRIPTION, Default: 0}

\item{nchar}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
