% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils4ggplot.r
\name{va_resc}
\alias{va_resc}
\title{2軸グラフ用の数値調整関数}
\usage{
va_resc(v, lim1, lim2, scale_which = c("var", "axis")[1])
}
\arguments{
\item{v}{Values to plot. 描画したい数値のベクトル.}

\item{lim1}{Limits of 1st y-axis. 左軸に対してplotしたい値の範囲.}

\item{lim2}{Limits of 2nd y-axis. 右軸に対してplotしたい値の範囲.}

\item{scale_which}{左軸に合わせて数値を変換するか, 右軸の目盛りを変換するか.
Default: "var". 目盛り調整の際は "axis" を選ぶ. 詳しくは examples を参照.}
}
\value{
変換された数値
}
\description{
2軸グラフを描く時は左軸の数値にスケーリングし, 右軸の目盛りを調整する必要があります.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{
\url{https://blog.statsbeginner.net/entry/2020/09/08/181536}
\url{https://knknkn.hatenablog.com/entry/2019/02/27/204913"}
\url{https://notchained.hatenablog.com/entry/2016/10/02/011204"}
}
